feat: cohort layout suspsense
