+++ apps/web/src/app/(dashboard)/(management)/locatie/[location]/cohorten/[cohort]/(overview)/_components/layout-tabs.tsx
@@ -1,75 +1,66 @@
-"use client";
-
-import type { enums } from "@nawadi/lib";
-import { clsx } from "clsx";
 import NextLink from "next/link";
-import { useParams, useSelectedLayoutSegments } from "next/navigation";
-import type { z } from "zod";
-import type { ActorType } from "~/lib/nwd";
+import { notFound } from "next/navigation";
+import { Suspense } from "react";
+import {
+  listPrivilegesForCohort,
+  listRolesForLocation,
+  retrieveCohortByHandle,
+  retrieveLocationByHandle,
+} from "~/lib/nwd";
+import { LayoutTabsClient } from "./layout-tabs-client";
 
-export function LayoutTabs({
-  personRoles,
-  personPrivileges,
-}: {
-  personRoles: ActorType[];
-  personPrivileges: z.infer<typeof enums.Privilege>[];
-}) {
-  const params = useParams();
-  const segments = useSelectedLayoutSegments();
+type Props = { params: Promise<{ location: string; cohort: string }> };
 
-  const hasRole = (role: ActorType[]) =>
-    personRoles.some((r) => role.includes(r));
+async function LayoutTabsContent(props: Props) {
+  const params = await props.params;
+  const location = await retrieveLocationByHandle(params.location);
+  const cohort = await retrieveCohortByHandle(params.cohort, location.id);
 
-  const hasPrivilege = (privilege: z.infer<typeof enums.Privilege>[]) =>
-    personPrivileges.some((p) => privilege.includes(p));
+  if (!cohort) {
+    notFound();
+  }
 
+  const [roles, privileges] = await Promise.all([
+    listRolesForLocation(location.id),
+    listPrivilegesForCohort(cohort.id),
+  ]);
+
+  return <LayoutTabsClient personRoles={roles} personPrivileges={privileges} />;
+}
+
+function LayoutTabsFallback() {
   return (
-    <>
-      <nav
-        className="flex space-x-6 text-sm overflow-auto scrollbar-none"
-        aria-label="Tabs"
-      >
-        {[
-          {
-            name: "Cursisten",
-            href: `/locatie/${params.location}/cohorten/${params.cohort}`,
-            enabled: hasRole(["instructor", "location_admin"]),
-            current: segments.length < 1,
-          },
-          {
-            name: "Diploma's",
-            href: `/locatie/${params.location}/cohorten/${params.cohort}/diplomas`,
-            enabled:
-              hasRole(["location_admin"]) ||
-              hasPrivilege(["manage_cohort_certificate"]),
-            current: segments[0] === "diplomas",
-          },
-          {
-            name: "Instructeurs",
-            href: `/locatie/${params.location}/cohorten/${params.cohort}/instructeurs`,
-            enabled:
-              hasRole(["location_admin"]) ||
-              hasPrivilege(["manage_cohort_students"]),
-            current: segments[0] === "instructeurs",
-          },
-        ]
-          .filter((tab) => !!tab.enabled)
-          .map((tab) => (
-            <NextLink
-              key={tab.name}
-              href={tab.href}
-              className={clsx(
-                tab.current
-                  ? "bg-slate-100 text-slate-700"
-                  : "text-slate-500 hover:text-slate-700",
-                "rounded-md px-3 py-2 text-sm font-medium",
-              )}
-              aria-current={tab.current ? "page" : undefined}
-            >
-              {tab.name}
-            </NextLink>
-          ))}
-      </nav>
-    </>
+    <nav
+      className="flex space-x-6 overflow-auto text-sm scrollbar-none"
+      aria-label="Tabs"
+    >
+      {[
+        {
+          name: "Cursisten",
+        },
+        {
+          name: "Diploma's",
+        },
+        {
+          name: "Instructeurs",
+        },
+      ].map((tab) => (
+        <NextLink
+          key={tab.name}
+          href="#"
+          className="px-3 py-2 rounded-md font-medium text-slate-500 hover:text-slate-700 text-sm"
+        >
+          {tab.name}
+        </NextLink>
+      ))}
+    </nav>
+  );
+}
+
+export function LayoutTabs(props: Props) {
+  return (
+    <Suspense fallback={<LayoutTabsFallback />}>
+      <LayoutTabsContent {...props} />
+    </Suspense>
   );
 }
